// types/chargeback.ts
import { BaseCreateRequestAttributes, BaseListParams, Relationship, Sort, Tags } from "./common"

export interface Chargeback {
    /**
     * Identifier of the chargeback resource.
     */
    id: string

    /**
     * Type of the chargeback resource. The value is always chargeback.
     */
    type: "chargeback"

    /**
     * JSON object representing the chargeback data.
     */
    attributes: {
        /**
         * The date the chargeback was created.
         * RFC3339 format. For more information: https://en.wikipedia.org/wiki/ISO_8601#RFCs
         */
        createdAt: string

        /**
         * The amount (in cents) of the chargeback.
         */
        amount: number

        /**
         * 	Description of the chargeback.
         */
        description: string

        /**
         * See [Tags](https://developers.unit.co/#tags).
         */
        tags?: object
    }

    /**
     * Describes relationships between the chargeback resource and other resources (accounts, transaction, customer).
     */
    relationships: {
        /**
         * The [Deposit Account](https://docs.unit.co/deposit-accounts/) the funds will be debited from.
         */
        account: Relationship

        /**
         * The [Customer](https://developers.unit.co/customers/) the deposit account belongs to.
         */
        customer: Relationship

        /**
         * The account that will receive the funds.
         */
        counterpartyAccount: Relationship

        /**
         * The Chargeback [Transaction](https://docs.unit.co/resources/#transaction-chargeback) generated by this chargeback.
         */
        transaction: Relationship
    }
}

export interface ChargebackListParams extends BaseListParams {
    /**
     * Optional. Filters the results by the specified account id.
     * default: empty
     */
    accountId?: string

    /**
     * Optional. Filters the results by the specified customer id.
     * default: empty
     */
    customerId?: string

    /**
     * Optional. Filter Chargebacks by [Tags](https://docs.unit.co/#tags).
     * default: empty
     */
    tags?: Tags

    /**
     * Optional. Filters the Chargebacks that occurred after the specified date. e.g. 2020-01-13T16:01:19.346Z
     */
    since?: string

    /**
     * Optional. Filters the Chargebacks that occurred before the specified date. e.g. 2020-01-02T20:06:23.486Z
     */
    until?: string

    /**
     * Optional. Leave empty or provide sort=createdAt for ascending order. Provide sort=-createdAt (leading minus sign) for descending order.
     * default: sort=-createdAt
     */
    sort?: Sort

    /**
     * Optional. A comma-separated list of related resources to include in the response. Related resources include: customer, account, transaction. See [Getting Related Resources](https://docs.unit.co/about-jsonapi/#intro-getting-related-resources).
     */
    include?: string
}

export interface CreateChargebackRequest {
    /**
     * Type of the resource to be created, the value is always chargeback.
     */
    type: "chargeback"
    
    attributes: {
        /**
         * The amount (in cents) to charge the account and credit the counterparty.
         */
        amount: number

        /**
         * Description of the chargeback (maximum of 50 characters).
         */
        description: string
    } & BaseCreateRequestAttributes

    relationships: {
        /**
         * The Deposit Account the funds will be debited from.
         */
        account: Relationship

        /**
         * The account that will receive the funds.
         */
        counterpartyAccount: Relationship
    }
}